backend/DAPPER/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
environs>=9.5.0
django-cors-headers>=4.0.0
openai>=1.0.0  # For OpenRouter API (using OpenAI client)
python-dotenv>=1.0.0
whitenoise>=6.5.0  # For serving static files in production
